feat: separate contents from files
