+++ src/workers/webdav/modules/contents/infrastructure/CachedRemoteFileContentsManagersFactory.ts
@@ -1,12 +1,12 @@
-import { ContentFileClonner } from '../../domain/ContentFileClonner';
-import { ContentFileDownloader } from '../../domain/ContentFileDownloader';
-import { ContentFileUploader } from '../../domain/ContentFileUploader';
-import { FileSize } from '../../domain/FileSize';
-import { RemoteFileContentsManagersFactory } from '../../domain/RemoteFileContentsManagersFactory';
-import { File } from '../../domain/File';
-import { ContentsCacheRepository } from '../../domain/ContentsCacheRepository';
+import { FileSize } from '../../files/domain/FileSize';
+import { ContentFileClonner } from '../domain/ContentFileClonner';
+import { ContentFileDownloader } from '../domain/ContentFileDownloader';
+import { ContentFileUploader } from '../domain/ContentFileUploader';
+import { ContentsCacheRepository } from '../domain/ContentsCacheRepository';
+import { RemoteFileContentsManagersFactory } from '../domain/RemoteFileContentsManagersFactory';
 import { CachedContentFileDownloader } from './download/CachedContentFileDownloader';
 import { CachedContentFileUploader } from './upload/CachedContentFileUploader';
+import { File } from '../../files/domain/File';
 
 export class CachedRemoteFileContentsManagersFactory
   implements RemoteFileContentsManagersFactory
