+++ src/workers/webdav/modules/files/test/__mocks__/RemoteFileContentsManagersFactoryMock.ts
@@ -1,9 +1,9 @@
 import { Readable } from 'stream';
-import { RemoteFileContentsManagersFactory } from '../../domain/RemoteFileContentsManagersFactory';
+import { RemoteFileContentsManagersFactory } from '../../../contents/domain/RemoteFileContentsManagersFactory';
 import { FileSize } from '../../domain/FileSize';
 import { File } from '../../domain/File';
 import { ContentFileUploader } from '../../domain/ContentFileUploader';
-import { ContentFileClonner } from '../../domain/ContentFileClonner';
+import { ContentFileClonner } from '../../../contents/domain/ContentFileClonner';
 import { ContentFileUploaderMock } from './ContentFileUploaderMock';
 import { ContentFileClonnerMock } from './ContentFileClonnerMock';
 import { ContentFileDownloaderMock } from './ContentFileDownloaderMock';
